week_2: подключил бд, настроил маппинги
- подключил mapstruct
- подключил mapstruct-binding
- настроил загрузку аннотаций lambok и mapstruct

+++ src/main/java/ru/skypro/diplom/dto/ads/CreateAdsDto.java
@@ -7,6 +7,6 @@ public class CreateAdsDto {
     private String description;
     private String image;
     private long pk;
-    private long price;
+    private int price;
     private String title;
 }
